c2fff57 Merge topic 'improved-ribbon-filter' into master
50eab6c Use vtkColor in the protein ribbon filter
5abadb7 Merge branch 'pv-master'
166d57c Improve the PDBReader to support secondary structures
1162574 Merge topic 'add_test_label_argument_for_vtk_module' into master
34278f3 Add extra test label argument for vtk_module macro
1e63e05 Merge topic 'Fix_NamedColors_Comparison' into master
bdb5a86 Merge topic 'enhance-plyreader-with-textures-and-normals' into master
2c0819a TestUCharToDouble() is failing on one machine.
33301f8 Merge topic 'Fix_TestNamedColors_warning' into master
83d1162 Fixed an unused variable warning.
711ebe8 Merge topic 'fix-warnings' into master
277684b Fix shadowed variable warning in vtkPriorityQueue
ad5bc81 Fix warning in vtkInteractiveChartXYZ
d3d0a59 Fix warning in vtkChartXYZ
e8e9db4 Fix "expression always true" warning in vtkHardwareSelector
...

+++ IO/Exodus/vtkExodusIIReader.cxx
@@ -4544,7 +4544,7 @@ void vtkExodusIIReaderPrivate::SetInitialObjectStatus( int objectType, const cha
     {
     idx += 4;
     idlen = 0;
-    while(idx+idlen < nm.length() &&  nm.at(idx+idlen) != ' ' )
+    while(idx+idlen < static_cast<int>(nm.length()) &&  nm.at(idx+idlen) != ' ' )
       {
       idlen++;
       }
